// src/clients/ts/types/resource.ts
// @generated by protobuf-ts 2.9.4
// @generated from protobuf file "types/resource.proto" (package "vink.kubevm.io.apis.types", syntax proto3)
// tslint:disable
/**
 * @generated from protobuf enum vink.kubevm.io.apis.types.ResourceType
 */
export enum ResourceType {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: VIRTUAL_MACHINE = 1;
     */
    VIRTUAL_MACHINE = 1,
    /**
     * @generated from protobuf enum value: VIRTUAL_MACHINE_INSTANCE = 2;
     */
    VIRTUAL_MACHINE_INSTANCE = 2,
    /**
     * @generated from protobuf enum value: DATA_VOLUME = 3;
     */
    DATA_VOLUME = 3,
    /**
     * @generated from protobuf enum value: NODE = 4;
     */
    NODE = 4,
    /**
     * @generated from protobuf enum value: NAMESPACE = 5;
     */
    NAMESPACE = 5,
    /**
     * @generated from protobuf enum value: MULTUS = 6;
     */
    MULTUS = 6,
    /**
     * @generated from protobuf enum value: SUBNET = 7;
     */
    SUBNET = 7,
    /**
     * @generated from protobuf enum value: VPC = 8;
     */
    VPC = 8,
    /**
     * @generated from protobuf enum value: IPPOOL = 9;
     */
    IPPOOL = 9,
    /**
     * @generated from protobuf enum value: STORAGE_CLASS = 10;
     */
    STORAGE_CLASS = 10,
    /**
     * @generated from protobuf enum value: IPS = 11;
     */
    IPS = 11,
    /**
     * @generated from protobuf enum value: VIRTUAL_MACHINE_SUMMARY = 12;
     */
    VIRTUAL_MACHINE_SUMMARY = 12,
    /**
     * @generated from protobuf enum value: EVENT = 13;
     */
    EVENT = 13
}